LogIn/LogOut
Code in _Layout that calls a static property in the TokenChecker class to determine if the user is logged in or not,

If the user has recieved a statuscode OK whilst trying to log in (recieved a token) the prop will get its value set to True.  Be aware this executed in the LoginView code-behind.

A new page is created: LogOut. If the link log out is clicked in _LayOut the propertyvalue will be set to false and the session will be cleared.

A test is conducted and you cant reach the admin-page after clicking logout.